fix: move ImageProcessor protocol to domain layer (fixes #85)
- Create ImageProcessor protocol in domain/repositories/image_processor.py
- Update imports in generation_service.py to use domain protocol
- Update infrastructure to import protocol from domain
- Update test imports to use domain protocol
- Add ImageProcessor to domain repositories exports

This fixes the DDD violation where domain layer was depending on
infrastructure layer. Now the dependency flow is correct:
domain -> defines protocol, infrastructure -> implements protocol.
