+++ src/emojismith/domain/repositories/__init__.py
@@ -1,5 +1,6 @@
 """Domain repositories."""
 
 from .slack_repository import SlackRepository
+from .image_processor import ImageProcessor
 
-__all__ = ["SlackRepository"]
+__all__ = ["SlackRepository", "ImageProcessor"]
